+++ dftd3/utils.py
@@ -28,18 +28,28 @@
 #
 
 
-from math import exp, sqrt
+from dataclasses import InitVar, dataclass, field
+from typing import List
 
 import jax.numpy as jnp
-from qcelemental import periodictable
 
 from .constants import AU_TO_ANG
-from .parameters import RCOV
+from .parameters import BJ_PARMS, RCOV, ZERO_PARMS
 
 
-def E_to_index(element):
-    """Convert element symbol to 0-based index."""
-    return periodictable.to_Z(element) - 1
+def der_order(order: int) -> str:
+    """Return correct suffix for order.
+
+    >>> der_order(1)
+    ... "1-st"
+    >>> der_order(4)
+    ... "4-th"
+    """
+    _suffix = ["st", "nd", "rd"]
+    if order - 1 > 3:
+        return f"{order}-th"
+    else:
+        return f"{order}-{_suffix[order - 1]}"
 
 
 def getMollist(bondmatrix, startatom):
@@ -167,3 +177,67 @@ def check_inputs(*, charges, coordinates):
         raise RuntimeError(
             f"The size of the coordinates (3*{len(coordinates) // 3}) and atom types ({natom}) arrays do not match!"
         )
+
+
+@dataclass
+class D3Configuration:
+    functional: str
+    damp: str = "zero"
+    s6: float = field(init=False)
+    rs6: float = field(init=False)
+    s8: float = field(init=False)
+    a1: float = field(init=False)
+    a2: float = field(init=False)
+    threebody: bool = False
+    bond_index: List[List[int]] = None
+    intermolecular: bool = False
+    pairwise: bool = False
+
+    # initialization-only variables
+    _s6: InitVar[float] = 0.0
+    _rs6: InitVar[float] = 0.0
+    _s8: InitVar[float] = 0.0
+    _a1: InitVar[float] = 0.0
+    _a2: InitVar[float] = 0.0
+
+    def __post_init__(self, _s6, _rs6, _s8, _a1, _a2):
+        which = self.damp if self.damp == "zero" else "Becke-Johson"
+        cfg = f">>> D3-dispersion correction with {which} damping\n"
+
+        if self.damp.casefold() == "zero".casefold():
+            if any(map(lambda x: x == 0.0, [_s6, _rs6, _s8])):
+                self.s6, self.rs6, self.s8 = ZERO_PARMS[self.functional]
+                where = "from database"
+            else:
+                self.s6 = _s6
+                self.rs6 = _rs6
+                self.s8 = _s8
+                where = "user-defined"
+            cfg += f"    - Damping parameters ({where}): s6 = {self.s6}; rs6 = {self.rs6}; s8 = {self.s8}\n"
+            # initialize remaining parameters
+            self.a1 = _a1
+            self.a2 = _a2
+        elif self.damp.casefold() == "bj".casefold():
+            if any(map(lambda x: x == 0.0, [_s6, _s8, _a1, _a2])):
+                self.s6, self.a1, self.s8, self.a2 = BJ_PARMS[self.functional]
+                where = "from database"
+            else:
+                self.s6 = _s6
+                self.a1 = _a1
+                self.s8 = _s8
+                self.a2 = _a2
+                where = "user-defined"
+            cfg += f"    - Damping parameters: s6 = {self.s6}; s8 = {self.s8}; a1 = {self.a1}; a2 = {self.a2}\n"
+            # initialize remaining parameters
+            self.rs6 = _rs6
+        else:
+            raise RuntimeError(f"{self.damp} is an unknown damping scheme.")
+
+        if not self.threebody:
+            cfg += "    - 3-body term will not be calculated\n"
+        else:
+            cfg += "    - Including the Axilrod-Teller-Muto 3-body dispersion term\n"
+        if self.intermolecular:
+            cfg += "    - Only computing intermolecular dispersion interactions! This is not the total D3-correction\n"
+
+        print(cfg)
